src/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
databricks-sql-connector==3.5.0
databricks-sdk==0.35.0
pydantic==2.9.2
